Add a minimal gasp table to all fonts
I removed it previously since it wasn't ClearType-ready and `psfnormalize` kept bugging me to remove it.
However, now FontBakery complains and suggests adding a dummy table with all four bits on for all sizes, so let's do that (although I fail to grasp why any modern rasterizer wouldn't just treat the absense of this table in the same way as presense of this dummy).