Increase descender sizes
Original fonts specified a line gap, which I removed during UFO normalization.
This caused a decrease in the overall line height, which was especially noticeable when switching between the old and new font in diffenator proof pages.
This change adds missing units to the descender, which restores the original line height and means that capitals are once again almost perfectly vertically centered within the line.